control: display file rotation properties (Bug: 459483)
Reviewed-on: https://git.eclipse.org/r/41535
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>

+++ org.eclipse.tracecompass.lttng2.control.core.tests/src/org/eclipse/tracecompass/lttng2/control/core/tests/model/impl/ChannelInfoTest.java
@@ -1,5 +1,5 @@
 /**********************************************************************
- * Copyright (c) 2012, 2014 Ericsson
+ * Copyright (c) 2012, 2015 Ericsson
  *
  * All rights reserved. This program and the accompanying materials are
  * made available under the terms of the Eclipse Public License v1.0 which
@@ -192,7 +192,7 @@ public class ChannelInfoTest {
         long subBufferSize = fixture.getSubBufferSize();
         assertEquals(1L, subBufferSize);
 
-        int maxSizeTraceFiles = fixture.getMaxSizeTraceFiles();
+        long maxSizeTraceFiles = fixture.getMaxSizeTraceFiles();
         assertEquals(1024, maxSizeTraceFiles);
 
         int maxNumberTraceFiles = fixture.getMaxNumberTraceFiles();
