+++ org.eclipse.tracecompass.lttng2.control.ui/src/org/eclipse/tracecompass/internal/lttng2/control/ui/views/service/LTTngControlServiceMI.java
@@ -446,7 +446,7 @@ public class LTTngControlServiceMI extends LTTngControlService {
                                 channel.setOutputType(attribute.getTextContent());
                                 break;
                             case MIStrings.TRACEFILE_SIZE:
-                                channel.setMaxSizeTraceFiles(Integer.parseInt(attribute.getTextContent()));
+                                channel.setMaxSizeTraceFiles(Long.parseLong(attribute.getTextContent()));
                                 break;
                             case MIStrings.TRACEFILE_COUNT:
                                 channel.setMaxNumberTraceFiles(Integer.parseInt(attribute.getTextContent()));
